correcting review and little refactor
