+++ src/frontend/src/logic/errors.js
