+++ src/frontend/src/views/LandingPage.vue
@@ -1,23 +1,23 @@
 <template>
   <div class="landingPage">
-    <Landing1/>
-    <Landing2/>
+    <LandingWelcome/>
+    <LandingHowTo/>
     <RecentAnimalsList/>
     <DevTeam/>
   </div>
 </template>
 
 <script>
-import Landing1 from '../components/Landing_1.vue'
-import Landing2 from '../components/Landing_2.vue'
-import RecentAnimalsList from '../components/RecentAnimalsList.vue'
-import DevTeam from '../components/Dev_team.vue'
+import LandingWelcome from '../components/landing/LandingWelcome.vue'
+import LandingHowTo from '../components/landing/LandingHowTo.vue'
+import RecentAnimalsList from '../components/landing/RecentAnimalsList.vue'
+import DevTeam from '../components/landing/DevTeam.vue'
 
 export default {
   name: 'App',
   components: {
-    Landing1,
-    Landing2,
+    LandingWelcome,
+    LandingHowTo,
     RecentAnimalsList,
     DevTeam
   }
